add testing data registry, move common testing functions to testing.helpers
