+++ src/ravenpy/testing/registry.txt
@@ -1,0 +1,58 @@
+basinmaker/drainage_region_0175_v2-1/finalcat_info_v2-1.zip sha256:a761d4a2284d728ff1c846db84d22dbb37afb3acc293f8b3110a56a99bc283c9
+caspar_eccc_hindcasts/geps_watershed.nc sha256:a7eddbdca773b0c66140fbfe042f1bb889a3cb1f368c0393d0598be58bea5208
+cec_nalcms2010_30m/cec_nalcms_subQC.tiff sha256:31143499c1368b7e7d73a5f33b9184aabb2eb869337352f0905853d39eb5675f
+clrh/mattawin/06FB002.rvh sha256:9ab58a048a358b159f681e90fbb129084232ba58a467d148b8b0e7f154222204
+clrh/mattawin/Lakes.rvh sha256:5d39d2dbaa96089b38952436583da76ba70a6ab587da5cadcb8d9034a684c7ef
+clrh/mattawin/channel_properties.rvp sha256:6b48f275e0fc5f1f323346502fb9a11f6c8869620f1ab89d9c24a6fef7d4633a
+donneesqc_mrc_poly/mrc_subset.gml sha256:89a7a47a008a04b5821d6fbd9d11a84b163b341e23798e2375bc0cf7cd29043e
+donneesqc_mrc_poly/mrc_subset.zip sha256:021bebe8abdea2ca0f4c605d6d38981d46bf4ec0c642dadf1be3aa34522cfd0d
+earthenv_dem_90m/earthenv_dem90_southernQuebec.tiff sha256:310f38da0439300f22751de434a9dbcff23d20376db19aff4e260573b526beff
+eccc_forecasts/geps_watershed.nc sha256:f63cab6543e86bdd6b96599b16cd5fecf9da2e022dd08b6351b6852fec598d21
+famine/famine_input.nc sha256:0026f20c141dc007f9041d1964e831c767a72206eb3690ce1ae328d940e1e6a4
+gr4j_cemaneige/solution.rvc sha256:99b25947c22a99ccdd5a738923ab17af100cee8d61c3e73969e07ee2472b457c
+hydro_simulations/raven-gr4j-cemaneige-sim_hmets-0_Hydrographs.nc sha256:a1763cb78ab81cae8080792cf99c52a42b0dff1f72f6b7b4862c421811599371
+matapedia/Matapedia_meteo_data_2D.nc sha256:1c4a10f7d7964e7943838c9b7725723a80742597bb0753ea53f6fd996f79c0c0
+matapedia/Matapedia_meteo_data_stations.nc sha256:20892444705b504598cadb816694626a93d46a7edac89f6068bc4e476b21e451
+matapedia/Qobs_Matapedia_01BD009.nc sha256:d94e6776954b4e4d05ce185889799b255db7e300f9e438129f53cfde6cb85b07
+nasa/Mars_MGS_MOLA_DEM_georeferenced_region_compressed.tiff sha256:3190c6799724b37c11c326db25cf3cebf57d87e0133e818e6aba18da91d43453
+nrcan/NRCAN_1971-1972_subset.nc sha256:7de9def968a7f3d804255be2b86976e47b0edc3a80e5bf4ad608c1adf2f01d40
+nrcan/NRCAN_2006-2007_subset.nc sha256:a029261f1b74cd688e0b7f70ab3a770e9e9e4f5f005b3a370707803550c1c646
+polygons/Basin_10.zip sha256:d611ec4d082fc7a26fbcfbcd843c04c155e66f2bc044896374218a19af4fc6d9
+polygons/Saskatoon.geojson sha256:5277ccb26affb3d9a7697d2b605fd0367b5621627f409ea17003b36c735e762b
+polygons/mars.geojson sha256:4eb7c1d20f43cf41d886026dd6b4601d087d352fc8988d5f43fb44402d07af14
+polygons/mars.zip sha256:504c509f4d7e6a6344e23a76831cba78491017b168f7c4b28752a3ad13d1f8cc
+polygons/mississippi.geojson sha256:089445dc9db5c93d851a2a654523871fae65996dd6123b40bfc4b0cbef32916e
+raven-gr4j-cemaneige/Salmon-River-Near-Prince-George_meteo_daily.nc sha256:68d234ae5a14051f8d0849ab204aa31c8d65dab4eaec65764306a23540cd8e9d
+raven-gr4j-cemaneige/Salmon-River-Near-Prince-George_meteo_daily_2d.nc sha256:2171a33280e7d2deff582bc2f7bb903a6dd520848be0c09495ae938f7c214f61
+raven-gr4j-cemaneige/Salmon-River-Near-Prince-George_meteo_daily_3d.nc sha256:fb61ca9fa69b3c2ccb56af9dd92211bed72aa386ef2aa941013ed2df238b21b7
+raven-gr4j-cemaneige/raven-gr4j-salmon.rvc sha256:465e617460e41b1ca80423a7722f080374407e5aa120be2121ab6a8f84b58bf6
+raven-gr4j-cemaneige/raven-gr4j-salmon.rvh sha256:8695bcc319ff833999e7914315e7cbc07edcc4d18e6be451ff570db23619b89b
+raven-gr4j-cemaneige/raven-gr4j-salmon.rvi sha256:1c1e087a746fab027cf11d348760422fdcb589e9fb70703fd2fb218781d21f35
+raven-gr4j-cemaneige/raven-gr4j-salmon.rvp sha256:a066e32951dd6cf248e753f2850e7146539f777f501de9f4f1a33def97ff4120
+raven-gr4j-cemaneige/raven-gr4j-salmon.rvt sha256:b0a6bf6835d8d43830f8477ca09c5542622d0db4d0f9e3eb6f663527f5606ef3
+raven-hmets/Salmon-River-Near-Prince-George_Qobs_daily.rvt sha256:ad4a92d6eb48f9501bc10bd88626135d8018392ecbeb99908e9f64ce76467d90
+raven-hmets/Salmon-River-Near-Prince-George_meteo_daily.rvt sha256:e1b47934873caa9b9ac44c8d3069e73520432b5e245124c255b3f5c52f940c17
+raven-hmets/raven-hmets-salmon.rvc sha256:8a79de8c19ff43a21688f494d1e5c6cb07227e8a83b9560c72b934db56eaa17f
+raven-hmets/raven-hmets-salmon.rvh sha256:b602fc0e523ef87888a0985f06b27c6d391ef7bfcef94f1104832b2a25b59ac4
+raven-hmets/raven-hmets-salmon.rvi sha256:f40e4db30a2569790cabc435568ee92fb528166fe95f5265988b3608de3327d2
+raven-hmets/raven-hmets-salmon.rvp sha256:4066a35d2638adab8b608a0721e41ed620f9ab9d7aa0837134b51aee560e3835
+raven-hmets/raven-hmets-salmon.rvt sha256:a32125fc1557cd4d460205a42cea0761203febf3cc7bc6ca9e4c0447292df018
+raven-mohyse/Salmon-River-Near-Prince-George_Qobs_daily.rvt sha256:ad4a92d6eb48f9501bc10bd88626135d8018392ecbeb99908e9f64ce76467d90
+raven-mohyse/Salmon-River-Near-Prince-George_meteo_daily.rvt sha256:e1b47934873caa9b9ac44c8d3069e73520432b5e245124c255b3f5c52f940c17
+raven-mohyse/raven-mohyse-salmon.rvc sha256:282a3ad02ffecb16f3f9ee5c4357a809254499039a4ab51d6b537a2ca9f2224f
+raven-mohyse/raven-mohyse-salmon.rvh sha256:80373ec8a80ee1aba9e8e5c04ea68d1f44de5fd693706dc5ebdeec6783548151
+raven-mohyse/raven-mohyse-salmon.rvi sha256:f6414972f4d170f9aa8532f5dcab45518bbd6938b2e2e952631381e05c6bcf1b
+raven-mohyse/raven-mohyse-salmon.rvp sha256:6f6440be3205407974479c9f7a36527fcb3973e074e3b8ab314a12d13befa7b1
+raven-mohyse/raven-mohyse-salmon.rvt sha256:4420c6db1f981b283fd82d8227efa8a88604ae01140afb35350b2a0fe9e3ab18
+raven-routing-sample/OTT_sub.zip sha256:f90de4ad5467dae7bcd061156c147cd09ec7d9b3182ac86cefaa1816a541db9b
+raven-routing-sample/VIC_streaminputs.nc sha256:ffea0e6b1095eabb417e048d3cc807d7964036ffda3e9cbe478dadb20908d7e9
+raven-routing-sample/VIC_streaminputs_weights.rvt sha256:2912c3604a5461dca27255bbdca5d319afd91a4534257391264a14f8c95bffb8
+raven-routing-sample/VIC_temperatures.nc sha256:d9109f115d12f15a4c311374b5770df3a1fd0f8b861bf9dc029ee90e497b9e8a
+raven-routing-sample/VIC_test_nodata.nc sha256:434cb2dbe3796ff08ff3d26189fade6d978eea91014cbae9c7df3ba6a949bfe8
+raven-routing-sample/VIC_test_nodata_weights.rvt sha256:bf430e9b9a29bb9c31bda87495e5c72822d32ec6cc6ba311b01997f32bcb4815
+raven-routing-sample/WSC02LE024.nc sha256:e76aa242b06eae78c6a9980592550abd154496b447e78a50167cfb2de0e8c41a
+raven-routing-sample/era5-test-dataset-crop.nc sha256:9e3088282022372c4737b04d3360ea88982d749d234ca24ca07743e0f6b28bde
+raven-routing-sample/finalcat_hru_info.zip sha256:dd6818455d9e967d000d4fbc3f33c5b68af9d1babe75b093e2bb29847acaf59a
+raven-routing-sample/lievre_hrus_v21.zip sha256:2dc6d6ab21f5b009e437da4324438d3cea3934ca0cfd72a0ed69a1deb3c8b6e3
+watershed_vector/Basin_test.zip sha256:c5c02f7fe8be37c22813715cdf30e03eb163113a6610d862ce280a2203895a7e
+watershed_vector/LSJ_LL.zip sha256:036d4a607a6d1f11eb4e73915be68fcad11477fecc927742f5b53d21d105bc5b
